refactor(staking): move more code to common. Delete unused.
