+++ packages/bzx-common/src/domain/ProviderTypeDetails.ts
