+++ packages/staking-dashboard/src/App/ProviderMenu/ProviderMenuListItem.tsx
@@ -1,6 +1,6 @@
 import { observer } from 'mobx-react'
 import React from 'react'
-import ProviderType from 'src/domain/ProviderType'
+import ProviderType from 'bzx-common/src/domain/ProviderType'
 import ProviderTypeDictionary from 'src/domain/ProviderTypeDictionary'
 import { Loader, ButtonBasic } from 'ui-framework'
 import Web3Connection from 'src/stores/Web3Connection'
