--- packages/staking-dashboard/src/domain/ClaimRequest.ts
@@ -1,8 +1,0 @@
-export default class ClaimRequest {
-  public id: number
-  public name: string
-  constructor() {
-    this.id = Math.round(new Date().getTime() / 1000)
-    this.name = 'Claim'
-  }
-}
