--- packages/staking-dashboard/src/domain/IWeb3ProviderSettings.ts
@@ -1,5 +1,0 @@
-export default interface IWeb3ProviderSettings {
-  networkId: number
-  networkName: string
-  etherscanURL: string
-}
