--- packages/staking-dashboard/src/domain/RequestStatus.ts
@@ -1,9 +1,0 @@
-enum RequestStatus {
-  AWAITING = 'Awaiting',
-  IN_PROGRESS = 'In progress',
-  FAILED = 'Failed',
-  FAILED_SKIPGAS = 'Failed_SkipGas',
-  DONE = 'Done'
-}
-
-export default RequestStatus
