+++ packages/staking-dashboard/src/domain/Web3ConnectionFactory.ts
@@ -1,13 +1,10 @@
-import {
-  RPCSubprovider,
-  Web3ProviderEngine
-} from '@0x/subproviders'
+import { RPCSubprovider, Web3ProviderEngine } from '@0x/subproviders'
 import { Web3Wrapper } from '@0x/web3-wrapper'
 import { AbstractConnector } from '@web3-react/abstract-connector'
 import { ConnectorUpdate } from '@web3-react/types'
 import configProviders from '../config/providers.json'
 import appConfig from 'src/config/appConfig'
-import ProviderType from './ProviderType'
+import ProviderType from 'bzx-common/src/domain/ProviderType'
 
 export default class Web3ConnectionFactory {
   public static rpcSubprovider: RPCSubprovider
